cleaning existing codes
