+++ Instantiate.xcodeproj/xcshareddata/xcschemes/Instantiate.xcscheme
@@ -1,39 +1,11 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <Scheme
-   LastUpgradeVersion = "9999"
+   LastUpgradeVersion = "0900"
    version = "1.3">
    <BuildAction
       parallelizeBuildables = "YES"
       buildImplicitDependencies = "YES">
       <BuildActionEntries>
-         <BuildActionEntry
-            buildForTesting = "YES"
-            buildForRunning = "YES"
-            buildForProfiling = "YES"
-            buildForArchiving = "YES"
-            buildForAnalyzing = "YES">
-            <BuildableReference
-               BuildableIdentifier = "primary"
-               BlueprintIdentifier = "OBJ_27"
-               BuildableName = "InstantiateStandard.framework"
-               BlueprintName = "InstantiateStandard"
-               ReferencedContainer = "container:Instantiate.xcodeproj">
-            </BuildableReference>
-         </BuildActionEntry>
-         <BuildActionEntry
-            buildForTesting = "YES"
-            buildForRunning = "YES"
-            buildForProfiling = "YES"
-            buildForArchiving = "YES"
-            buildForAnalyzing = "YES">
-            <BuildableReference
-               BuildableIdentifier = "primary"
-               BlueprintIdentifier = "OBJ_37"
-               BuildableName = "InstantiateTestsResource.framework"
-               BlueprintName = "InstantiateTestsResource"
-               ReferencedContainer = "container:Instantiate.xcodeproj">
-            </BuildableReference>
-         </BuildActionEntry>
          <BuildActionEntry
             buildForTesting = "YES"
             buildForRunning = "YES"
@@ -67,6 +39,15 @@
             </BuildableReference>
          </TestableReference>
       </Testables>
+      <MacroExpansion>
+         <BuildableReference
+            BuildableIdentifier = "primary"
+            BlueprintIdentifier = "OBJ_36"
+            BuildableName = "Instantiate.framework"
+            BlueprintName = "Instantiate"
+            ReferencedContainer = "container:Instantiate.xcodeproj">
+         </BuildableReference>
+      </MacroExpansion>
       <AdditionalOptions>
       </AdditionalOptions>
    </TestAction>
@@ -83,9 +64,9 @@
       <MacroExpansion>
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "OBJ_27"
-            BuildableName = "InstantiateStandard.framework"
-            BlueprintName = "InstantiateStandard"
+            BlueprintIdentifier = "OBJ_36"
+            BuildableName = "Instantiate.framework"
+            BlueprintName = "Instantiate"
             ReferencedContainer = "container:Instantiate.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
@@ -98,6 +79,15 @@
       savedToolIdentifier = ""
       useCustomWorkingDirectory = "NO"
       debugDocumentVersioning = "YES">
+      <MacroExpansion>
+         <BuildableReference
+            BuildableIdentifier = "primary"
+            BlueprintIdentifier = "OBJ_36"
+            BuildableName = "Instantiate.framework"
+            BlueprintName = "Instantiate"
+            ReferencedContainer = "container:Instantiate.xcodeproj">
+         </BuildableReference>
+      </MacroExpansion>
    </ProfileAction>
    <AnalyzeAction
       buildConfiguration = "Debug">
